refactor: rename MultivariateAnalysisForm files
