+++ src/features/multivariate_layer/helpers/openMultivariateModal.ts
@@ -1,5 +1,5 @@
 import { showModal } from '~core/modal';
-import { MultivariateAnalysisForm } from '../components/MultivariateAnalysisForm';
+import { MultivariateAnalysisForm } from '../components/MultivariateAnalysisForm/MultivariateAnalysisForm';
 import type { MultivariateLayerConfig } from '~core/logical_layers/renderers/MultivariateRenderer/types';
 
 export async function openMultivariateModal(initialConfig?: MultivariateLayerConfig) {
